为了避免误导，直接禁止KEYS模糊搜索

+++ XUnitTest/SortedSetTests.cs
@@ -22,7 +22,7 @@ namespace XUnitTest
             var file = @"config\redis.config";
             if (File.Exists(file)) config = File.ReadAllText(file.GetFullPath())?.Trim();
             if (config.IsNullOrEmpty()) config = "server=127.0.0.1:6379;db=3";
-            File.WriteAllText(file.GetFullPath(), config);
+            if (!File.Exists(file)) File.WriteAllText(file.GetFullPath(), config);
 
             _redis = new FullRedis();
             _redis.Init(config);
